+++ .vscode/solution-explorer/class.cs-template
@@ -1,0 +1,8 @@
+using System;
+
+namespace {{namespace}}
+{
+    public class {{name}}
+    {
+    }
+}
