+++ .vscode/solution-explorer/class.ts-template
@@ -1,0 +1,3 @@
+export class {{name}} {
+
+}
