+++ .vscode/solution-explorer/class.vb-template
@@ -1,0 +1,9 @@
+Imports System
+
+Namespace {{namespace}}
+
+    Public Class {{name}}
+    
+    End Class
+
+End Namespace
