+++ .vscode/solution-explorer/default.ts-template
@@ -1,0 +1,3 @@
+export default {{name}} {
+
+}
