+++ .vscode/solution-explorer/enum.cs-template
@@ -1,0 +1,8 @@
+using System;
+
+namespace {{namespace}}
+{
+    public enum {{name}}
+    {
+    }
+}
