+++ .vscode/solution-explorer/interface.cs-template
@@ -1,0 +1,8 @@
+using System;
+
+namespace {{namespace}}
+{
+    public interface {{name}}
+    {
+    }
+}
