+++ .vscode/solution-explorer/interface.ts-template
@@ -1,0 +1,3 @@
+export interface {{name}} {
+
+}
