+++ 1.Chapter.Fundamentals/1.1.Chapter.BaseProgramModel/Ex_1.1.15/Ex_1.1.15.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>Ex_1._1._15</RootNamespace>
+  </PropertyGroup>
+
+</Project>
