+++ 1.Chapter.Fundamentals/1.1.Chapter.BaseProgramModel/Ex_1.1.24/Ex_1.1.24.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>_1._1.Chapter</RootNamespace>
+    <RootNamespace>Ex_1._1._24</RootNamespace>
   </PropertyGroup>
 
 </Project>
